+++ Makefile.am
@@ -3623,6 +3623,7 @@ libnss_sss_la_SOURCES = \
     src/sss_client/nss_services.c \
     src/sss_client/sss_cli.h \
     src/sss_client/nss_compat.h \
+    src/sss_client/nss_common.h \
     src/sss_client/nss_mc_common.c \
     src/util/io.c \
     src/util/murmurhash3.c \
